requirements.txt
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
flake8==6.1.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
ipaddress==1.0.23
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyflakes==3.1.0
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0